Fix new players implementation
Player is a class with safe instance. It means that you can copy
it with default copy constructor with minimal costs and only one
instance of its data will live until last object is deleted.
But when it comes to pointers and references to this object
things are getting little complicated, so I came back with
old fashioned pointers to Players and things are now cleaner and
better.